Semantics/recursion (#243)
* Rules to process `Ref` and `Deref` `RValues`

Evaluation of rules for `Local` assignment of references and dereferences to other `Local`

* Added simple test

* Simple semantics for `CopyForDeref(Deref)` supported

* Added test for double reference and double dereference

* Internal panic called for recurisve function call

* Naive recurision supported

Identifies the function name and reproduces the cell. This is not optimised as there is a lot of duplicated data. Furthermore there is currently no clearing of the data to an initial state when the frame is duplicated.

* Updated `compiletest-run-fail.tsv`

* Added fibonacci test

* Added mutual recursion test

* Fixed copy paste error from test runner

* fixed whitespace and removed failing proof (will be fixed later)