Added evaluation of `Option<...>::unwrap()`
Uses internal function `#unwrap(...)` as an executed `Statement` before executing next basic block.

Passes 6 more tests, however many failing tests have changed from `4` to `113`. I believe this is not actually because of this commit, but because this commit has allowed the test to get further to run into other issues.