Parser usability and bug fixes (#401)
* Added information to parser notes.

* Changed parser notes to markdown.

* Added testing path that uses kparse.

This enables writing tests without having to write the AST directly.
Therefore, it is easier to write larger tests.

* Removed unused testing of deprecated convert.py script.

* Additional tests.

* Changed symbols for lists in body.k according to parser requirements.

* Added missing files for previously added tests.

* Bug fix (list arguments were reversed), and additional tests.

* Make parser use sort instead of symbol as input.

* Additional tests - BasicBlock.

* Changed __main__.py argument name and description from symbol to sort.

* Removed change from symbol input to sort from parser notes.md.

* Bug fix, with additional tests. The bug was exposed with the test for operand.

Single sort enum productions with no named fields were handled wrongly.
The parser assumed that the detected dictionary was the enum's named
arguments, when it actually was the single, unnamed argument.

* Additional tests.

* Bug fix, with test that triggered the bug.

The issue was not including None in the potential types of Json
input to be handled. The rest of the code was already handling None.

* Added comments in parser.

* Removed commented out rules from ty.k

I will include in the PR description.

* Set Version: 0.3.31

* Removed commented out rules from kmir.k.

---------

Co-authored-by: devops <devops@runtimeverification.com>