`--start-symbol` for `kmir run` (#446)
This adds a `<start-symbol>` cell to the configuration with a
configuration variable for the symbol to begin execution from.

It also changes the expected output in the `exec-smir` integration tests
to only check the contents of the K cell. The semantics are going to be
very unstable right now, and without this change we will probably be
spending time updating the test vectors on nearly every pull request.

---------

Co-authored-by: devops <devops@runtimeverification.com>