Merge `kmir/Makefile` into `./Makefile`, adapting the CI workflow (#451)
All targets from `kmir/Makefile` were ported (i.e., just copied,
actually) into the top-level `Makefile`.

The new default target is `build`, which installs the python code and
builds the K definition.

The `test-integration` target (previously in `kmir/Makefile`) will now
check that the K definition is up to date before testing, to avoid a
stale `definition.kore`. `test-unit` does not, in order for CI to not
require a K framework to run it.

Running the global `make test` will execute both the python (unit and
integration) tests and the `smir-parse-tests`. The latter still do _not_
depend on `build` because we want to run them in CI with a
docker-supplied `K` toolkit and `pyk` without running the entire test
within docker (because it also needs a Rust toolchain).

---------

Co-authored-by: devops <devops@runtimeverification.com>