Match on sorted entries instead of use lookups + projections (#459)
This changes a few rules that use lookups on collections to instead
match on them in a more structural manner.

These changes are all in `kmir.md`, there's one more in there that's a
little more involved which could use more investigation. There are also
lots more in `rt/data.md`, but I have another PR open that would create
a lot of conflicts with changes made there, so I'm waiting to do that.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Jost Berthold <jost.berthold@gmail.com>