457 - kmir prove (#472)
fixes: #457 

This adds the `kmir prove` command which can load `.k` files containing
specification modules and run the prover on the claims contained within
them. It has support for `--proof-dir` to save the KCFGs as well as
`--include-labels` and `--exclude-labels` to specify which claims to run
if you don't want to run all of them.

---------

Co-authored-by: devops <devops@runtimeverification.com>