Load proof from proof dir when viewing
The previous code was creating a new proof object from the spec file (first claim).
Instead, the proof needs to be loaded from a file in the proof directory.

Note that `APRProver.read_proof` and `APRProver/read_proof_data` have different expectations to how the proof is stored. The latter method has to be used.