Update `kmir run` to behave more like `cargo run` (#524)
- Introduce `CargoProject` class which can be used to gather information
about a cargo project.
- `kmir run` now uses `CargoProject` which finds the `.smir.json`
artifact in the target directory after invoking `cargo build`.

Note that currently either `RUSTC` will need to be set to the
`stable-mir-json` install, or the smir json files should already have
been built that way and exist in the target directory. Later, we can get
rid of this requirement when we package `stable-mir-json` with `kmir`.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Jost Berthold <jost.berthold@gmail.com>