Parser fixes/improvements (#534)
This makes a few bug fixes for the symbol names and group attributes in
the semantics that didn't line up with what was emitted by
stable-mir-json. It also makes a some performance improvements by
caching a few repeated computations.

There are a couple of changes included here that are in a slightly
larger scope:

- `null` bytes in byte arrays from stable-mir appear to be
uninitialized. We currently don't have a way to handle uninitialized
bytes so we default to 0 when we see these.
- The semantics for reserving non-primitive symbolic values no longer
need the type for them in the type map. This change was made because the
type information we would have needed wasn't in the type map from
stable-mir-json. There are changes coming down the pipe that include the
information and we can revert this later.

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>
Co-authored-by: devops <devops@runtimeverification.com>