Migrate to using uv for build system (#537)
We're working on migrating our build systems to use `uv` instead of
`poetry`, as it's a more modern/faster/shinier thing.

See here for more reasoning:
https://github.com/runtimeverification/kontrol/pull/1011

This largely follows that PR for making the migration, but skips all the
Nix flake as my understanding is none of that is working in this repo at
the moment anyway.

In addition, the `Dockerfile` setup is adjusted to not run as `root` on
CI, to make sure it's more similar to an actual user's environment.

---------

Co-authored-by: devops <devops@runtimeverification.com>