Adding logging and pruning and proof resumption (#539)
Needed verbosity and debug logging in the typical RV way. It is
important to add the flag for verbosity in the right place, which is
after the first kmir command, e.g.:

```
poetry run -C kmir/ kmir run --verbose <FILE>
poetry run -C kmir/ kmir prove --verbose run <FILE>
```

Also added the ability to prune and resume proofs for better debugging.

IN PARTICULAR:

- Option `--max-depth` to `kmir prove` allows for many steps the prover
will take between nodes
- Option `--max-iterations` to `kmir prove` allows to restrict the
number of steps the prover can take in total
- `kmir prove` will now check if the proof exists on disc, if it does it
will use this instead of restarting. Use `--reload` to force it to
restart anyway.
- `kmir prune --proof-dir <DIR> <proof-id> <node_id>` is added to remove
a node and subsequent nodes in the graph from the proof saved on disc.

---------

Co-authored-by: devops <devops@runtimeverification.com>