Inline `Tools` class into `KMIR` and replumb (#549)
In preparation of building LHSs of specs directly, I've inlined the
`Tools` class directly into the `KMIR` class, and propagated all the
appropriate changes through.

This also required getting `KParse` to match the diamond import pattern
of `KRun` and `KProve` (which I think is eventually the intention, to
upstream it), so I did that as well.

Overall, this moves `make_init_config` and `run_parsed` directly into
`KMIR`, so it's available for all commands that `KMIR` exposes.

As one minor build system improvement, I allow the user to set variable
`PARALLEL` on the CLI when invoking `make` to spede things up.

---------

Co-authored-by: devops <devops@runtimeverification.com>