Faster rerun on already passing proofs (#553)
This PR makes sure that we are not starting up the RPC server unless
it's actually needed, and also not calling cargo again unless actually
needed.

In particular:

- The call to `cargo ...` which produces SMIR for building a proof only
runs if we are actually in the branch which will use it in
`KMIR.prove_rs`.
- If the proof we load from disk is already passing, we don't bother
starting up the `KCFGExplore` (which starts up the booster server).
- Some formatting cleanups.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Jost Berthold <jost.berthold@gmail.com>