Unchecked arithmetic proofs and lemmas (#571)
The proofs for unchecked arithmetic (from AWS initiative challenge 11)
require additional lemmas to confirm that no truncation is happening.
These lemmas are added here, also removing one lemma that we had twice
(in the lemmas file and in `numbers.md` where `truncate` is defined).

Also checking in the Rust code over which we run the proofs, and will
set up a convention for which functions to run in `prove-rs`
(suggestion: one-line comment containing a list of `start-symbol`
parameters at the top of the file).

---------

Co-authored-by: devops <devops@runtimeverification.com>