Adding machine info (#575)
The `MachineInfo` semantics productions are now connected to the JSON
which now includes it with an update. This is ignored in the `Pgm`
currently, but could be used to make `isize` and `usize` dynamic on
`pointer-width` - although, I am not sure how to do that atm.

This is currently pointing to a different branch of stable mir json

---------

Co-authored-by: devops <devops@runtimeverification.com>