Add tests for mutability (#574)
Adding tests to examine the boundaries of mutability and immutability,
not all of these use `unsafe` but only the ones I listed as unsafe do we
annotate unsafe directly:
- mut_const.rs: A UB test that mutates the data under an immutable value
through pointer arithmetic
- interior-mut-fail.rs: Safely has interior mutability through `RefCell`
(stuck on `Rvalue::AddressOf`)
- interior-mut2-fail.rs: Safely has interior mutability through `Cell`
(stuck on `Rvalue::AddressOf`)
- interior-mut3-fail.rs: Unsafe but sound interior mutability through
`UnsafeCell` (stuck on `Rvalue::AddressOf`)

We discussed the possibility of adding tests that show demonstrate
immutable field of mutable structs, but this is not possible in rust.
There is a notion of private fields if structs come from other modules,
but this is not the same as mutability.

---------

Co-authored-by: devops <devops@runtimeverification.com>