Remove mutability checks on lookup path of #projectedUpdate (#582)
We currently are checking, _nearly_ every step of the way when looking
up a location to write to with `#projectedUpdate`, the mutability of
that location. But we actually only should need to check the mutability
of the final location we are writing to. This PR removes those interior
mutability checks.

---------

Co-authored-by: devops <devops@runtimeverification.com>