Added `thunk` constructor (#569)
Following the conversation where we decided that we would create a
constructor to delay evaluation of currently impossible to implement
features (e.g. when an address of a place is required) Everett and I
have implemented the `thunk(Evaluation)` constructor to `TypedValue`. We
have a `thunk` first approach, meaning that the errors are removed and
left as `thunk`s. This should stop us getting stuck and allow us to get
very far in our executions with unevaluated `thunk`s that will
illustrate features we implement.

Things added:
- `TypedValue ::= thunk(Evalution)`

Things changed:
- `Evaluation ::= #cast`
- `Evaluation ::= #compute`
- `Evaluation ::= #expect`

Removed rules resulting in:
- MIRIndexError
- MIRConstantIndexError
- UnexpectedCastArgument
- UnexpectedCastTarget
- UnknownCastTarget
- CastNotImplemented

Because we have not implemented the logic of calculating
`tyOfLocal(...)` or `mutabilityOf(...)` for the `thunk`, there are many
places where we now basically forget the type or mutability of a given
location. This is OK, because it makes our execution more conservative.
If that information is needed to continue execution, we'll get stuck
now, and so we won't accidentally prove something that's not true.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>
Co-authored-by: Jost Berthold <jost.berthold@gmail.com>