Bitwise operations (#579)
This PR implements implements bitwise operators and fixing a bug
uncovered doing so.

## Bitwise Operators

Bitwise operators `&` "And", `|` "Or", `^` "Xor":
- `&`, `|`, `^` on integers
- `&`, `|`, `^` on booleans
- `&`, `|`, `^` on borrows (Not completed, stuck on `GlobalAllocs` /
promoteds)
both LHS and RHS of the operation must have the same type.

Shifts `<<` and `>>` are also implemented, although shifts are broken
down into these commands:
- `Shl` shift left with wrapping on overflow
- `Shr` arithmetic shift right with wrapping on overflow
- `ShlUnchecked` shift left with UB on overflow
- `ShrUnchecked` arithmetic shift right with UB on overflow
where overflow is defined as the right hand arg (shift amount) being
greater than the number of bits in the left hand arg.

Concrete proofs are added for the above, borrows are currently failing.

## Bug Found and Fixed

This PR also uncovered a bug we had in how we were handling `Terminator`
`SwitchInt`. This value is _always_ interpreted as an unsigned value
(even if the branching is on a negative value). We were considering the
magnitude of the value, but needed to interpret the bits as unsigned.

---------

Co-authored-by: devops <devops@runtimeverification.com>