Initialise a call config in python (with symbolic arguments) (#584)
`make_call_config` produces a configuration with a `Call` terminator for
the given start symbol in the K cell.
Arguments to the function will be created as locals of a dummy caller's
stack frame, and copied by `OperandCopy` in the `Call` terminator.

Arguments are symbolic values, with as much structure as the type
metadata for the argument types allows.

* Primitive types are created with variables for the values. For
`Integer`, constraints limit the size of the value.
* `struct`s and `enum`s are created as `Aggregate` values but with a
single variable (list-sorted) for the arguments. With more metadata for
structs, we could create the struct fields with their resp. types.
* Tuples are created with a list of their components (recursively
structured)
* References are created by creating the pointee value _behind_ the
argument list

The dummy caller's locals have the following structure:

```
  Return    | Function arguments           | Pointees for
  value     | (symbolic structured values) | ref. arguments
------------------------------------------------------------
NewValue(0) | ARG_1 | ARG_2 | ... | ARG_N  | P_1 | P_2 | ...
-------------------------------------------|----------------
                       |                           A
                       \__(offset 0, place(N+2)____/
```
For sake of an example, ARG_2 here is a reference to data that was
(symbolically) created at location P_2, behind the locals that serve as
arguments of the called function. When the called function's locals are
set up, the reference in `ARG_2` is copied and its offset adjusted to 1.

---------

Co-authored-by: Guy Repta <50716988+gtrepta@users.noreply.github.com>
Co-authored-by: devops <devops@runtimeverification.com>