Improvements to printing using new infrastructure (#593)
These are some improvements to printing that we can implement now that
we have merged changes upstream
https://github.com/runtimeverification/k/pull/4808.

- Improves the printing of spans in the TUI viewer to actually print the
file name and line that it's referring to.
- Passes the entire `DisplayOpts` to the `KMIRAPRNodePrinter`, because
why not?
- Omits printing the `<currentBody>` cell by default in TUI and shower,
but enables turning it back on with `--no-omit-current-body` option to
both.
- Saves off the `SMIRInfo` in the `proof_dir` when we have it available
in `prove-rs`, so that it can be used for printing sourcemap information
in the show/view commands (like current function name and span).

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Jost Berthold <jost.berthold@gmail.com>