Refactor reading/writing projections to deduplicate logic (#601)
Currently, we have duplicated logic about how to traverse projections
between the `#readProjection` and `#projectedUpdate`. This PR goes about
deduplicating this logic, and some re-arrangements of the code
afterwards. This PR has several big changes in separate commits, it may
be easier to review commit-by-commit rather than all at once.

In particular:

- The `#projectedUpdate` had fields `UPDATE` (for the new value), and
`FORCE` (for whether to force setting even an immutable local), which
were not used until the very end when it got turned back into a
`#setLocalValue`. They're removed from `#projectedUpdate` and put into a
second `KItem` on the `<k>` cell instead (in an item called
`#writeProjection`), so that they don't need to show up in the rules.
- The cases `operandMove` and `operandCopy` are updated to use
`#projectedUpdate` to traverse the projectiosn directly, and to hold
whether it's a copy/move in the second item of the `<k>`-cell in a
`#readProjection`. Once the projcetions are done being traversed, we
have logic for correctly marking the slot as moved (or not), and then
returning the looked up value (and ignoring the generated context).
- All the old rules for `#readProjection` are deleted (which were
largely duplicate with `#projectedUpdate`).
- `#projectedUpdate` is renamed to `#traverseProjection`.
- Old helpers and such which were located next to teh `#readProjection`
rules are now moved to be next to the relevant `#traverseProjection`
rules.
- The operations `operandMove` and `operandCopy` used to handle the
cases with empty `.ProjectionElems` differently (by directly doing the
move/copy). Those special cases are eliminated, so taht it goes through
all the same traversal logic (which has the same effect in the end, but
reduces the number of potential rules to try).
- Formatting is updated.
- Test output is updated.

---------

Co-authored-by: devops <devops@runtimeverification.com>