Refactor SMIRInfo Usage (#602)
Closes: https://github.com/runtimeverification/mir-semantics/issues/603

This PR is several refactors to how SMIRInfo is built/used to make it
easier to do a later refactor I'm doing to unstick an example that was
supplied. It also includes some formatting changes that were in my local
work:

- Formatting:
- The rules for `#decodeConstant` are updated to include mention of the
`<k>` cell (instead of implicit).
  - Some rule spacing/newline changes.
- Two test expected output for `exec-smir` leftover from before depth
changes were applied are deleted.
- Comments where were removed
https://github.com/runtimeverification/mir-semantics/pull/601 are added
back.
  - The file `data.md` is organized a little bit more.
- Places that were using `stringToken` are now using the more generic
`token` from the `pyk.kast.prelude`.
- `SMIRInfo` is a bit better encapsulated, and has some more processing
logic:
- `SMIRInfo.dump` is added, which can write the original SMIR JSON
output to a file.
- Places outside of `SMIRInfo` where we were handling the JSON directly
of `stable-mir-json` are now made to always go through the `SMIRInfo`
object instead. That includes in `parse/parser.py` and in
`KMIR.prove_rs`.
- Some warnings are added for cases where we cannot find the needed
function symbols/definitions in the smir.

These changes make it a bit easier to make the next logical changes I
have to make, which are related to handling intrinsic to get deeper into
execution.

---------

Co-authored-by: devops <devops@runtimeverification.com>