Python construction of initial states (#604)
This PR switches us over to fully building the initial configuration in
Python instead of in K. It leaves the K construction in place, in case
anyone wants to use the old kprove/krun tools to run it, but otherwise
mostly ignores it.

In particular:

- Populating configuration directly in Python:
- The cells `<functions>`, `<types>`, and `<adt-to-ty>` are populated
directly with maps instead of using K functions to construct the maps in
the first step.
- `KMIR.make_call_config` takes as input a `SMIRInfo` directly, and only
uses `Parser` for the small components it needs to build for making the
initial configuration.
- `KMIR.make_call_config` is enabled to make concrete configurations
where the cell contents mimics that of the `configuration` declaration.
- `KMIR.run_parsed` is renamed to `KMIR.run_smir`, and takes as input a
`SMIRInfo` and uses `KMIR.make_call_config` instead of
`KMIR.make_init_config` to build the initial state.
- Unused functions `KMIR.make_init_config` and `KMIR.run_call` are
removed.
- Minor testing updates:
- Updating the expected output is sped up by skipping tests that do not
have expected output when that flag is passed, and ensuring that we
update expected output even if pass/fail status of test changes.
- The `exec-smir` test-suite expected files have the state depth removed
from the file name.
- The depths for the `exec-smir` tests are adapted to the new setup
(increasing depth by 1).
- Expected output is updated (some change to `smir-exec` test output as
we're calling a function directly instead now).
- Adds two tests, one from the Kani test-suite about reading outside the
bounds of a given location, and one as a way of using checked arithmetic
to assert the precondition on unchecked arithmetic. Both are failing
(the out of bounds read _should_ be failing, the unchecked arithmetic
should not).

---------

Co-authored-by: devops <devops@runtimeverification.com>