Fixes to linker output, dead item elimination (#610)
* Added empty spans array (temporary) and machine info to linked output
The spans will have to be offset just like `Ty` and `AdtDef` IDs to be
in the linked ouput (for later).

* Added methods to remove all items unreachable from a given start
symbol from the `SMIRInfo` (creating a copy)
* Using the reduced `SMIRInfo` for `kmir run` and `kmir prove-rs`.

In the process of implementing this, I also removed special handling of
the `main` symbol. We are now inserting the correct `symbol_name` into
the `function_symbols` table (previously, it was just `main` which made
the algorithm delete it).