Build stable-mir-json in `make test-integration` (#612)
When updating the project and attempting to run the integration tests on
master I got an error in the python code constructing `SmirInfo`. The
reason is that while `make test-integration` builds a kmir, it does not
build `stable-mir-json`. It was confusing me since I had updated the
submodule and the hashes were the same, and I could see there was a
build and populated `.stable-mir-json` directory (it was stale).

This PR moves the building of `stable-mir-json` into the make target
`test-integration`. It should not increase performance cost if one is
already build as cargo will report the project successfully build in
`0.0` seconds and continue.

I modified the CI test suite accordingly.