615 use value sort for evaluation (#625)
Large refactoring to have `TypedValue` only in the `<locals>` list,
while operating only with `Value`s when we read, write, or compute with
them.
* `Evaluation` sort has result `Value` now
* Projection traversal and helper functions recurse and return `Value`
now
* A few places where type information is needed for processing (pointer
casts, for one) are now using a function that apply projections to type
metadata.
* The `Moved` marker was previously a `TypedValue` but is now just
another `Value`. This is in order to not lose type and mutability when
values are moved and the local is used again (typically in loops)

* I added some `Ceil` rules that appeared useful to avoid ceil
conditions on symbolic arguments.
* I had to add an `interim-simplification` option to the server to work
around a problem with `symbolic-args-fail.main`. Without adding that,
the server would grow an enormous memory footprint during the proof. I
experimented with the request in question but could not find out where
exactly the memory is allocated. It is _not_ Haskell heap data, so
probably the LLVM backend allocates it somehow.