Assign Ty IDs to all function items
The function table in SMIR JSON currently only contains functions which are called from other functions within the crate.
For library crates, there can be many functions in the crate that aren't iun the table, and for binary crates, `main` is usually missing.

Previously, only `main`'s symbol was added to the table.

This change assigns negative `Ty` IDs to every item which is not present in the function table.

As a result, then dead-item pruning code now works on all functions without crash.

The previous hack to expose start symbol functions was removed from the crate tests.