529 pointer cast corrections (#632)
This became a rather large and pervasive change, adding metadata to the
`Reference` value (not just to pointers).

* `Reference`s now carry metadata which gives the size of the pointee in
case of an array or slice.
* The size of slices (and arrays) is important when dereferencing the
references (or pointers, which is much less common).
* Statically-determined sizes are also carried around to avoid looking
up the size from the `TypeInfo` each time
* when dereferencing arrays, only the expected size is copied. If the
expected size is larger than the actual array, the execution gets stuck.
A new test has been added which is expected to fail on undefined
behaviour (`rustc`-compiled code will instead happily copy random data
into the array, as can be observed by activating the `println`s in the
new test).