Hotfix avoid deref in refs (#654)
Before creating a reference or pointer, the pointee is evaluated using
`#traverseProjection` and then the projections are reconstructed. This
shortens out all `Deref` projections, avoiding the problem of creating a
reference with offset `-1` (which can only happen when the returned
value refers to another local reference or pointer and is dereferenced
immediately).