Backport p-token tweaks (#672)
This is a collection of changes that were (or appeared) useful while
working on the P-Token domain data model and advancing the property
proofs. Each commit is an addition of its own
* Ceil simplification rules for `#mapOffset` (for the legacy backend,
might not be necessary after all)
* avoid modulus calculations for number truncation (we now have a
function for it)
* project setup: add markdown selectors `symbolic`/`concrete` so we can
separate symbolic backend code blocks
* implement `CastKindIntToInt` for `Bool` -> `Int` (related code was
observed in the wild).
* Implement support for trivial cases of `transmute` (same type only)

The "avoid modulus" requires changes (good changes) on the test
expectations, the changed `Ceil` simplification changes a hash in one
expectation file.