HOTFIX: Avoid fall-back on call terminator rules
The current call terminator rules look up the function `MonoItemKind` as a match in
the function map using the map key `#tyOfCall(FUNC)`. This is not supported by booster
and leads to fall-backs on every function call.

The change avoids the map match by moving the key computation and lookup into the
side condition in both rules (intrinsic and non-intrinsic rule). Step count is unchanged.