Publish Docker images for feature/p-token branch (#700)
A new workflow is added which publishes a docker image for kmir from the
`feature/p-token` on every push.

The workflow is largely copied from the `release.yml` workflow, only the
docker image tag is different.

* For release: using `ubuntu-jammy-` and the `package/version` string
* Here: using `p-token-` and the short commit SHA

This image will be used from the solana-token fork, at the SHA of the
submodule dependency (which should always be a commit from the
`feature/p-token` branch).